Fix mocked `check_env` environment tree
The environment tree of a mocked exercise checking environment now more closely resembles the environment chain produced by `gradethis_exercise_checker()`.

In the future we might want to revisit `mock_this_exercise()` and refactor it to use `preapre_check_env()` directly, but since `mock_this_exercise()` has to do some of the work we expect to be handled by learnr, it might not be worth it.

Fixes #270